Move  uploadarea patcher to packageinstaller
